requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
APScheduler==3.10.4
aiofiles==23.2.1
openai==1.12.0    